Restructuring of measurement and backlog sharding in prep to InfluxDB v2
Implementation of both auth modalities. Basic and Token.

It now aims at the /api/v2/write, which means no support for 1.x
anymore.

Measurement does not contain the retention policy, precision etc. This
is due of them not being part of the API they where in InfluxDBv1 or
being part of the request to API itself instead of the line.
We now have Measurement be a 1:1 mapping to a line in a write request.

Backlog gets now plugged into Client via Trait constraint instead of
wrapping around it. This makes the Client API more accessible by
avoiding the need of importing it via Trait first.

File Backlog now shards data into one file per record. This way the
metadata is encoded in the file name (org_bucket_precision.log). This
allows for backlog processing being able to map 1:1 with the Client API
flow. One file to one record to one write request.
To a lesser degree we also get limited sharding by Record. And it opens
up the possibility to further shard by size to accomodate for the 5k
lines per request optimum mentioned in InfluxDB documentation.

Closes #1, #2, #8